Expand character set from 80 to 64K characters
Each character is now addressed with 2 bytes instead of 1. The number of
bytes per character is configured in (I think) exactly one spot, so
increasing or decreasing this is trivial. English speakers can just set
it to 1.

The animator seems a little unstable; if I leave my character in a
public for a while, the board becomes unresponsive. Oh well.

* Check in fonts. Did this so users don't have to remember to set the
  resolution or to disable mipmaps.