Fix matchStrings O(n^2) loop
This slides 2 windows across input strings, looking for a region where
they are most similar. It then uses that region to stitch the strings
together. Since transcribe.py passes in a continuous transcription as
the `old_text` argument, we can wind up spending a lot of time here.

Constrain the area of the `old_text` argument that we look at to the
most recent 50 characters. This should be good enough.

Also fix how we calculate levenshtein_distance. Uh... yeah, let's not
talk about how it was before.