Begin work on obfuscation
The basic idea is that we can raise the barrier to entry for potential
data miners by encrypting traffic with a pre-shared key. Any data miner
would probably have access to both the compiled shader and network data,
which is obviously sufficient to decrypt that data. But they would have
to spend a little time figuring it out, which should defeat most casual
miners.