Add emotes
Add emotes.py. It accepts a list of images and creates a texture with 64 total
embedded images.

The shader knows how to draw these into fixed 6-character-wide slots. Each slot
must be aligned to a 6-character boundary. osc_ctrl has to pad with spaces
to make this work.

This whole patch is a little more complicated than it has any right to be, but
my brain feels fuzzy and I don't know where to start fixing it, so I'm going to
leave it shitty-but-functional for now.

There's also some bug where writing a character into the 11th slot causes it to
show up at the end of the board. I'll figure that out later, idk.

I didn't include any of the emotes I use since I couldn't find any info on
their licenses. I'm just banking on having a good workflow later on so
people can add their own.