Bugfix: don't use last region if it's partial
Because we allow users to customize the # of sync params, the board is
no longer divided into regions of uniform size. When the last region is
a different size than the rest, we simply omit it from paging.

This is a hack but it's easy to reason about.

Of course the entire paging stack should be rewritten, but not today.