Delete python310._pth
I was using this file to constrain the set of paths that Python can see,
but since `future` doesn't have a wheel, it will fail to install on a
fresh system.

If you set pip's --cache-dir to some new directory, you'll see it fail
to install.

The _pth doesn't really seem to matter, since without it, packages are
still installed under the virtual environment.