All transcription panel fields now persist across app restart
I forgor to put them into ApplyConfigToInputFields.

The reason this is necessary: we need to create the text field where we
log things before we can deserialize the config. To keep the code
structure "clean" I just wrote another function to apply the config
(ApplyConfigToInputFields). However I have to remember to update it when
I add new fields.