Begin rewriting transcribe.py
A set of proper interfaces is called for. See #dev-update-spam in
discord for drawing of design.

Also add code to mechanically optimize committer parameters using an
audio file. Not perfectly repeatable since it depends on the performance
characteristics of the machine, but prob better than what we had before
(nothing).